requirements.txt
mysql-connector==2.2.9
tkcalendar==1.6.1